refactor backend folders
